requirements.txt
selenium==4.15.0
requests==2.31.0
pytest==7.4.3
allure-pytest==2.13.2
python-dotenv==1.0.0
webdriver-manager==4.0.2